refactor: extract ValidationStepBase from Steps 7 and 8
Extract shared validation logic into ValidationStepBase abstract class:
- ReadGateMode (gate config parsing with fallbacks)
- DetermineSuccess (verdict-to-boolean mapping with warnings)
- CreateBlockingFailure (artifact failure creation)
- EnsureValidationDirectory (stale artifact cleanup)
- WriteSummarySection (markdown summary with append support)
- GetValidationDir (path resolution)

Both ArticleHealthValidatorStep and CoverageAuditStep now extend
ValidationStepBase instead of NamespaceStepBase directly, eliminating
~120 lines of duplicated logic. Step-specific behavior (script path,
prerequisite checks, article resolution) remains in each subclass.

No behavioral changes. All 448 tests pass.
